NTv2 .gsb file reader & interpolator
Python script reads the binary NTv2 file for AUSGeoid2020, finds the four surrounding nodes for the coordinates supplied and performs bilinear or bicubic interpolation (based on user input, or bicubic if not specified) . The output is a tuple of the ellipsoid to AHD separation, deflection in the prime meridian and deflection in the prime vertical values in that order (NAHD, ξ, η).

The script contains 2 classes with multiple functions, but only NTv2ReaderBinary().ntv2reader(file, lat, lon, interpolation_method) needs to be called, where:
• file is the complete file location as a string
• lat is the latitude of the point of interest in decimal degrees
• lon is the longitude of the point of interest in decimal degrees
• interpolation_method is either ‘bilinear’ or ‘bicubic’ (by default it is set to ‘bicubic’, so this parameter isn’t mandatory).

The script also contains 2 stand alone functions - one for bilinear and one for bicubic interpolation.

Signed-off-by: Jaimie Dodd <jaimie.dodd@dnrme.qld.gov.au>