Script written by Jeffrey Pickford (Senior Surveyor, QLD DNRME) with functions to manage calls to GeodePy functions to transform large amounts of data in csv format (such as the QLD SCDB) through the use of pandas.
Functions can be called either within a GUI or as stand alone functions.

grid2geoio() reads the csv file provided as a pandas dataframe and calls the grid2geo function in the geodePy transform module on the columns containing easting, northing and utmzone, and writes the results to new columns for latitude, longitude, point scale factor and grid convergence. Latitude and longitude can be in either DD or DMS depending on user input.

geo2gridio() reads the csv file provided as a pandas dataframe and calls the geo2grid function in the geodePy transform module on the columns containing latitude and longitude, and writes the results to new columns for easting, northing, utmzone, point scale factor and grid convergence. Latitude and longitude can be in either DD or DMS depending on user input.

gdatrans7() reads the csv file provided as a pandas dataframe and uses the Transformation class from the geodePy constants module, as well as calls to llh2xyz, conform7 and xyz2llh in order to transform coordinates from GDA2020 to GDA94 or from GDA94 to GDA2020 depending on the direction specified by the user.

Output from all functions is a new csv file containing all original data, as well as the columns created by the function.

Signed-off-by: Jaimie Dodd <jaimie.dodd@dnrme.qld.gov.au>